Move ProductCarousel to components directory
