+++ frontend/src/components/ProductCarousel.js
