+++ frontend/src/pages/Product.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
-import ProductCarousel from "../ProductCarousel";
+import ProductCarousel from "../components/ProductCarousel";
 import { MoneyFormatter } from "../localization";
 
 const Product = () => {
